Split CLI and fix datasets
